更新浏览次数

+++ longyu-common/src/main/java/com/longyu/common/domain/vo/ArticleDetailVo.java
@@ -48,6 +48,6 @@ public class ArticleDetailVo implements Serializable {
     /**
      * 访问量
      */
-    private Integer viewCount;
+    private Long viewCount;
 
 }
